+++ src/lantern/lib/metadata_library/models/record/presets/citation.py
@@ -34,8 +34,9 @@ def make_magic_citation(
     href = "?"
     try:
         self_identifier = next(identifier for identifier in identifiers if identifier.namespace == CATALOGUE_NAMESPACE)
-        href = self_identifier.href
-    except StopIteration:
+        if isinstance(self_identifier.href, str):
+            href = self_identifier.href.replace("lantern.", "")
+    except (StopIteration, TypeError):
         pass
     reference = f"[{href}]({href})"