+++ tests/resources/records/item_cat_verify.py
@@ -269,7 +269,7 @@ map_purchase = Distribution(
     ),
     transfer_option=TransferOption(
         online_resource=OnlineResource(
-            href="https://www.bas.ac.uk/data/our-data/maps/how-to-order-a-map/",
+            href="https://data.bas.ac.uk/guides/map-purchasing/",
             function=OnlineResourceFunctionCode.ORDER,
             title="Map ordering information - BAS public website",
             description="Access information on how to order item.",